SC update

+++ smart-contracts/contracts/interfaces/ICurveFiBase.sol
@@ -2,6 +2,6 @@ pragma solidity ^0.6.6;
 
 interface ICurveFiBase {
   function get_virtual_price() external view returns (uint);
-  function coins(int128 arg0) external view returns (address);
+  function coins(uint256 arg0) external view returns (address);
   function exchange(int128 from, int128 to, uint256 _from_amount, uint256 _min_to_amount) external;
 }
