+++ smart-contracts/contracts/only-for-tests/Curve3PoolForTest.sol
@@ -39,8 +39,8 @@ contract Curve3PoolForTest is ICurveFi3 {
         token = ERC20ForTest(_pool_token);
     }
     
-    function coins(int128 arg0) external view override returns (address) {
-        return i_coins[uint256(arg0)];
+    function coins(uint256 arg0) external view override returns (address) {
+        return i_coins[arg0];
     }
 
     function get_virtual_price() external view override returns (uint256) {
